+++ src/main/java/com/jdk/springrestmvc/service/FilmService.java
@@ -18,8 +18,8 @@ public interface FilmService {
     
     public void saveOrUpdate (Film film);
     
-    public void deleteFilm (int id);
+    public void deleteFilm (long id);
     
-    public Film findFilmById (int id);
+    public Film findFilmById (long id);
     
 }
